requirements.txt
amqp==2.5.2
antlr4-python3-runtime==4.8
arvados-cwl-runner>=2.1.0
arvados-python-client==2.0.2
asgiref==3.2.7
git+https://github.com/bio-ontology-research-group/bh20-seq-resource.git@master
billiard==3.6.3.0
bioblend>=0.14.0
CacheControl==0.11.7
celery==4.4.2
certifi==2020.4.5.1
CFGraph==0.2.1
chardet==3.0.4
ciso8601==2.1.3
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-recaptcha2==1.4.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
django-cors-headers>=3.5.0
future==0.18.2
google-api-python-client==1.6.7
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
isodate==0.6.0
jsonasobj==1.2.1
kombu==4.6.8
lockfile==0.12.2
mistune==0.8.4
oauth2client==4.1.3
oauthlib==3.1.0
pbr==5.4.5
pkg-resources==0.0.0
psycopg2==2.8.5
py-dateutil==2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.43.0.5
PyJSG==0.9.1
pyparsing==2.4.7
PyShEx==0.7.14
PyShExC==0.7.0
python-magic==0.4.18
python-memcached==1.59
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
rdflib==4.2.2
rdflib-jsonld==0.5.0
redis==3.5.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.15.77
ruamel.yaml.clib==0.2.0
schema-salad==7.0.20200612160654
ShExJSG==0.6.8
six==1.14.0
sparql-slurper==0.3.4
SPARQLWrapper==1.8.5
sqlparse==0.3.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
ws4py==0.5.1
zipp==3.1.0
